Fixed player being double damaged and int damage calculation
- removed Player tag check in Damage since the player already inherits from IDamage, and gets damaged because of it

- cast HP to a float in updatePlayerUI to work for HP and Stamina canvases

+++ Platforming-FPS/Assets/Scripts/Hiry/gameUIManager.cs
@@ -242,8 +242,8 @@ public class gameUIManager : MonoBehaviour
 
     public void updatePlayerUI()
     {
-        playersHealthPool.fillAmount = gameManager.instance.playerScript.HP / gameManager.instance.playerScript.HPOrig;
-        playersStaminaPool.fillAmount = gameManager.instance.playerScript.Stamina / gameManager.instance.playerScript.StaminaOrig;
+        playersHealthPool.fillAmount = (float)gameManager.instance.playerScript.HP / gameManager.instance.playerScript.HPOrig;
+        playersStaminaPool.fillAmount = (float)gameManager.instance.playerScript.Stamina / gameManager.instance.playerScript.StaminaOrig;
         gameUIManager.instance.RespawnCount.text = gameManager.instance.respawns.ToString("F0");
         ExperienceCount.text = gameManager.instance.playerScript.currentExperience.ToString("F0");
     }
